💄 style(agent-goals): move the goal mechanism explainer behind a hint
The "what happens after you create one" block sat inline on the goal empty
state, pushing the create button and the seeded examples down on every visit.
It now lives in a modal opened from a small "How it works" hint at the bottom,
laid out vertically since the three steps are a sequence.
