+++ src/features/AgentGoals/GoalEmptyState.tsx
@@ -3,17 +3,15 @@
 import { Block, Flexbox, Icon, Text } from '@lobehub/ui';
 import { Button } from '@lobehub/ui/base-ui';
 import { createStaticStyles, cssVar } from 'antd-style';
+import type { TargetIcon } from 'lucide-react';
 import {
   CalendarClockIcon,
   CheckIcon,
-  CircleCheckBigIcon,
+  CircleHelpIcon,
   InfinityIcon,
   Layers3Icon,
-  PlayIcon,
   PlusIcon,
-  RotateCcwIcon,
   TablePropertiesIcon,
-  TargetIcon,
   XIcon,
 } from 'lucide-react';
 import { memo } from 'react';
@@ -21,6 +19,7 @@ import { useTranslation } from 'react-i18next';
 
 import type { GoalExampleKey, GoalExampleSeed } from './goalExamples';
 import { buildGoalExampleSeed, GOAL_EXAMPLE_KEYS } from './goalExamples';
+import { createGoalHowItWorksModal } from './GoalHowItWorksModal';
 
 const styles = createStaticStyles(({ css }) => ({
   example: css`
@@ -80,67 +79,24 @@ const styles = createStaticStyles(({ css }) => ({
     max-width: 560px;
     line-height: 1.7;
   `,
+  howHint: css`
+    align-self: flex-start;
+    color: ${cssVar.colorTextTertiary};
+
+    &:hover {
+      color: ${cssVar.colorText};
+    }
+  `,
   judge: css`
     padding-block: 10px;
     padding-inline: 14px;
     border-radius: ${cssVar.borderRadius};
     background: ${cssVar.colorFillQuaternary};
   `,
-  loopBack: css`
-    padding-block: 10px;
-    padding-inline: 12px;
-    border: 1px dashed ${cssVar.colorBorderSecondary};
-    border-radius: ${cssVar.borderRadius};
-  `,
-  loopIcon: css`
-    flex: none;
-    margin-block-start: 1px;
-    color: ${cssVar.colorTextQuaternary};
-  `,
   section: css`
     padding-block: 24px;
     padding-inline: 40px;
     border-block-start: 1px solid ${cssVar.colorBorderSecondary};
-  `,
-  step: css`
-    min-width: 0;
-    height: 100%;
-    padding-block: 14px;
-    padding-inline: 16px;
-    border-radius: ${cssVar.borderRadius};
-
-    background: ${cssVar.colorFillQuaternary};
-  `,
-  stepHead: css`
-    /* Reserve a stable header height so the descriptions below start on the same
-       line across all three cards, even if one title wraps. */
-    min-height: 20px;
-  `,
-  stepIndex: css`
-    display: flex;
-    flex: none;
-    align-items: center;
-    justify-content: center;
-
-    width: 18px;
-    height: 18px;
-    border-radius: 999px;
-
-    font-size: 11px;
-    font-variant-numeric: tabular-nums;
-    color: ${cssVar.colorTextSecondary};
-
-    background: ${cssVar.colorFillSecondary};
-  `,
-  steps: css`
-    display: grid;
-    grid-template-columns: repeat(3, minmax(0, 1fr));
-    gap: 10px;
-    align-items: stretch;
-
-    @media (width <= 860px) {
-      grid-template-columns: minmax(0, 1fr);
-    }
   `,
 }));
 
@@ -149,30 +105,6 @@ const EXAMPLE_ICONS: Record<GoalExampleKey, typeof TargetIcon> = {
   digest: CalendarClockIcon,
   metric: Layers3Icon,
 };
-
-interface StepProps {
-  desc: string;
-  icon: typeof TargetIcon;
-  index: number;
-  title: string;
-}
-
-const Step = memo<StepProps>(({ desc, icon, index, title }) => (
-  <Flexbox className={styles.step} gap={8}>
-    <Flexbox horizontal align={'center'} className={styles.stepHead} gap={8}>
-      <span className={styles.stepIndex}>{index}</span>
-      <Icon icon={icon} size={14} style={{ flexShrink: 0 }} />
-      <Text fontSize={13} weight={600}>
-        {title}
-      </Text>
-    </Flexbox>
-    <Text fontSize={12} style={{ lineHeight: 1.65 }} type={'secondary'}>
-      {desc}
-    </Text>
-  </Flexbox>
-));
-
-Step.displayName = 'GoalEmptyStateStep';
 
 interface GoalEmptyStateProps {
   onCreate: (seed?: GoalExampleSeed) => void;
@@ -184,9 +116,10 @@ interface GoalEmptyStateProps {
  * A goal only pays off if the user understands the bargain before making one:
  * it runs autonomously, judges itself each round, and spends budget doing it.
  * "Goals will appear here" taught none of that, so this screen carries the
- * concept (what a goal is), the mechanism (what happens after you create one),
- * and three seeded examples that demonstrate what a *judgeable* outcome reads
- * like.
+ * concept (what a goal is) and three seeded examples that demonstrate what a
+ * *judgeable* outcome reads like. The mechanism — what happens round after
+ * round — is one level down, behind the hint at the bottom, so the two things
+ * that actually start a goal stay above the fold.
  */
 const GoalEmptyState = memo<GoalEmptyStateProps>(({ onCreate }) => {
   const { t } = useTranslation('chat');
@@ -209,38 +142,6 @@ const GoalEmptyState = memo<GoalEmptyStateProps>(({ onCreate }) => {
           <Button icon={PlusIcon} type={'primary'} onClick={() => onCreate()}>
             {t('goalEmpty.create')}
           </Button>
-        </Flexbox>
-      </Flexbox>
-
-      <Flexbox className={styles.section} gap={14}>
-        <Text fontSize={13} type={'secondary'} weight={600}>
-          {t('goalEmpty.howTitle')}
-        </Text>
-        <div className={styles.steps}>
-          <Step
-            desc={t('goalEmpty.step1.desc')}
-            icon={TargetIcon}
-            index={1}
-            title={t('goalEmpty.step1.title')}
-          />
-          <Step
-            desc={t('goalEmpty.step2.desc')}
-            icon={PlayIcon}
-            index={2}
-            title={t('goalEmpty.step2.title')}
-          />
-          <Step
-            desc={t('goalEmpty.step3.desc')}
-            icon={CircleCheckBigIcon}
-            index={3}
-            title={t('goalEmpty.step3.title')}
-          />
-        </div>
-        <Flexbox horizontal align={'flex-start'} className={styles.loopBack} gap={8}>
-          <Icon className={styles.loopIcon} icon={RotateCcwIcon} size={13} />
-          <Text fontSize={12} style={{ lineHeight: 1.6 }} type={'secondary'}>
-            {t('goalEmpty.loop')}
-          </Text>
         </Flexbox>
       </Flexbox>
 
@@ -310,6 +211,16 @@ const GoalEmptyState = memo<GoalEmptyStateProps>(({ onCreate }) => {
             <Text fontSize={12}>{t('goalEmpty.judge.good')}</Text>
           </Flexbox>
         </Flexbox>
+
+        <Button
+          className={styles.howHint}
+          icon={CircleHelpIcon}
+          size={'small'}
+          type={'text'}
+          onClick={() => createGoalHowItWorksModal()}
+        >
+          {t('goalEmpty.howHint')}
+        </Button>
       </Flexbox>
     </Block>
   );
